[feat]: organize utils and add pipelined recorder for saving data
